feat: score, timer 위치 및 애니메이션 교체
