+++ src/pages/games/SnackGame/game/ui/Score.ts
@@ -1,16 +1,16 @@
 import gsap from 'gsap';
-import { Container } from 'pixi.js';
+import { Container, Graphics } from 'pixi.js';
 
 import { Cloud } from './Cloud';
 import { Label } from './Label';
+import { Waves } from './Waves';
+import { app } from '../SnackGameBase';
 /**
  * 게임 플레이 동안 표시되는 게임 점수입니다.
  */
 export class Score extends Container {
   /** 애니메이션을 위한 내부 컨테이너 */
   private container: Container;
-  /** 애니메이션된 구름 배경 */
-  private cloud: Cloud;
   /** 표시되는 점수 숫자 */
   private messageLabel: Label;
   /** 현재 설정된 점수 */
@@ -19,31 +19,66 @@ export class Score extends Container {
   private showing = true;
   /** 실제 점수에 도달할 때까지 점진적으로 증가하는 점수 */
   private animatedPoints = 0;
+  /** 파도 효과 */
+  private readonly waves: Waves;
+  /** 애니메이션 컨테이너 원형 마스크 */
+  private readonly background: Graphics;
+  /** 배경 색상을 위한 그래픽 컴포넌트 */
+  private readonly circle: Graphics;
   /** 점수가 업데이트될 때마다 증가하는 빈도수에 따라 sfx 재생 피치 변경 */
   private intensity = 0;
 
   constructor() {
     super();
 
+    this.background = new Graphics();
+    this.background.circle(0, 0, app.renderer.height * 0.06);
+    this.background.fill(0xfff7ec);
+    this.addChild(this.background);
+
+    this.circle = new Graphics();
+    this.circle.circle(0, 0, app.renderer.height * 0.06);
+    this.circle.fill(0xffffff);
+    this.addChild(this.circle);
+
     this.container = new Container();
+    this.container.mask = this.circle;
     this.addChild(this.container);
 
-    this.cloud = new Cloud({
-      color: 0xfff7ec,
-      width: 200,
-      height: 20,
-      circleSize: 50,
-    });
-    this.container.addChild(this.cloud);
+    this.waves = new Waves([0xea4141, 0xff7979]);
+    this.waves.x = -100;
+    this.container.addChild(this.waves);
 
     this.messageLabel = new Label('0', {
       fill: 0xfb923c,
       fontSize: 30,
       fontFamily: 'DovemayoGothic',
     });
-    this.messageLabel.y = 8;
     this.container.addChild(this.messageLabel);
+
     this.points = 0;
+  }
+
+  // waves width 설정
+  public set width(value: number) {
+    this.waves.width = value;
+    this.waves.x = -value / 2;
+    this.waves.y = 40;
+  }
+
+  // waves hegith 설정
+  public set height(value: number) {
+    this.waves.height = value * 2;
+  }
+
+  /** waves 컴포넌트의 포지션 y값을 조금씩 뺍니다. */
+  public async upWavesPosition() {
+    console.log(this.waves.y);
+    gsap.to(this.waves, {
+      y: this.waves.y - 1,
+      duration: 1,
+      ease: 'back.in',
+    });
   }
 
   /** 점수를 0으로 리셋합니다 */
