+++ src/pages/games/SnackGame/game/ui/Timer.ts
@@ -1,76 +1,43 @@
 import gsap from 'gsap';
 import { Container, Graphics } from 'pixi.js';
 
+import { Cloud } from './Cloud';
 import { Label } from './Label';
-import { Waves } from './Waves';
-import { app } from '../SnackGameBase';
 
 /**
  * 게임 플레이 중에 표시되는 게임 타이머, 남은 시간이 10초 미만일 때 빨간색으로 깜빡이기 시작함.
  */
 export class Timer extends Container {
+  /** 애니메이션된 구름 배경 */
+  private cloud: Cloud;
   /** 애니메이션을 위한 컴포넌트 컨테이너 */
   private readonly container: Container;
   /** 표시되는 남은 시간 */
   private readonly messageLabel: Label;
-  /** 파도 효과 */
-  private readonly waves: Waves;
-  /** 애니메이션 컨테이너 원형 마스크 */
-  private readonly circle: Graphics;
-  /** 배경 색상을 위한 그래픽 컴포넌트 */
-  private readonly background: Graphics;
   /** 숨김시 false */
   private showing = true;
 
   constructor() {
     super();
 
-    this.background = new Graphics();
-    this.background.circle(0, 0, app.renderer.height * 0.06);
-    this.background.fill(0xfff7ec);
-    this.addChild(this.background);
-
-    this.circle = new Graphics();
-    this.circle.circle(0, 0, app.renderer.height * 0.06);
-    this.circle.fill(0xffffff);
-    this.addChild(this.circle);
-
     this.container = new Container();
-    this.container.mask = this.circle;
     this.addChild(this.container);
 
-    this.waves = new Waves([0xea4141, 0xff7979]);
-    this.waves.x = -100;
-    this.container.addChild(this.waves);
+    this.cloud = new Cloud({
+      color: 0xfff7ec,
+      width: 200,
+      height: 20,
+      circleSize: 50,
+    });
+    this.container.addChild(this.cloud);
 
     this.messageLabel = new Label('5:00', {
       fontSize: 32,
       fontFamily: 'DovemayoGothic',
       fill: 0xfb923c,
     });
+    this.messageLabel.y = 8;
     this.container.addChild(this.messageLabel);
-  }
-
-  // waves width 설정
-  public set width(value: number) {
-    this.waves.width = value;
-    this.waves.x = -value / 2;
-    this.waves.y = 50;
-  }
-
-  // waves hegith 설정
-  public set height(value: number) {
-    this.waves.height = value * 2;
-  }
-
-  /** waves 컴포넌트의 포지션 y값을 조금씩 뺍니다. */
-  public async upWavesPosition() {
-    if (this.waves.y < -50) return;
-    gsap.to(this.waves, {
-      y: this.waves.y - 1,
-      duration: 1,
-      ease: 'back.in',
-    });
   }
 
   /**
